Add a subject importer
This eliminates the need to enter subjects by hand. However, it still
requires us to go to the Academic Planner and copy/paste HTML into a
file. UM's site uses Ajax for the list of subjects, so it'd be a little
more involved to acquire automatically (something like capybara/webrat
might help).